@@ -1,7 +1,7 @@
 [package]
 name = "librespot-audio"
 version = "0.5.0-dev"
-rust-version = "1.61"
+rust-version = "1.73"
 authors = ["Paul Lietar <paul@lietar.net>"]
 description = "The audio fetching logic for librespot"
 license = "MIT"
@@ -19,7 +19,9 @@ bytes = "1"
 ctr = "0.9"
 futures-core = "0.3"
 futures-util = "0.3"
-hyper = { version = "0.14", features = ["client"] }
+hyper = { version = "1.3", features = [] }
+hyper-util = { version = "0.1", features = ["client"] }
+http-body-util = "0.1.1"
 log = "0.4"
 parking_lot = { version = "0.12", features = ["deadlock_detection"] }
 tempfile = "3"
