@@ -1,7 +1,7 @@
 [package]
 name = "librespot-core"
 version = "0.5.0-dev"
-rust-version = "1.61"
+rust-version = "1.73"
 authors = ["Paul Lietar <paul@lietar.net>"]
 build = "build.rs"
 description = "The core functionality provided by librespot"
@@ -15,7 +15,7 @@ version = "0.5.0-dev"
 
 [dependencies]
 aes = "0.8"
-base64 = "0.21"
+base64 = "0.22"
 byteorder = "1.4"
 bytes = "1"
 dns-sd = { version = "0.1", optional = true }
@@ -25,10 +25,12 @@ futures-util = { version = "0.3", features = ["alloc", "bilock", "sink", "unstab
 governor = { version = "0.6", default-features = false, features = ["std", "jitter"] }
 hmac = "0.12"
 httparse = "1.7"
-http = "0.2"
-hyper = { version = "0.14", features = ["client", "http1", "http2", "tcp"] }
-hyper-proxy = { version = "0.9", default-features = false, features = ["rustls"] }
-hyper-rustls = { version = "0.24", features = ["http2"] }
+http = "1.0"
+hyper = { version = "1.3", features = ["http1", "http2"] }
+hyper-util = { version = "0.1", features = ["client"] }
+http-body-util = "0.1.1"
+hyper-proxy2 = { version = "0.1", default-features = false, features = ["rustls"] }
+hyper-rustls = { version = "0.26", features = ["http2"] }
 log = "0.4"
 nonzero_ext = "0.3"
 num-bigint = { version = "0.4", features = ["rand"] }
@@ -38,9 +40,9 @@ num-traits = "0.2"
 once_cell = "1"
 parking_lot = { version = "0.12", features = ["deadlock_detection"] }
 pbkdf2 = { version = "0.12", default-features = false, features = ["hmac"] }
-priority-queue = "1.2"
+priority-queue = "2.0"
 protobuf = "3"
-quick-xml = { version = "0.31", features = ["serialize"] }
+quick-xml = { version = "0.32", features = ["serialize"] }
 rand = "0.8"
 rsa = "0.9.2"
 serde = { version = "1.0", features = ["derive"] }
