@@ -1,7 +1,7 @@
 [package]
 name = "librespot-discovery"
 version = "0.5.0-dev"
-rust-version = "1.61"
+rust-version = "1.73"
 authors = ["Paul Lietar <paul@lietar.net>"]
 description = "The discovery logic for librespot"
 license = "MIT"
@@ -10,7 +10,8 @@ edition = "2021"
 
 [dependencies]
 aes = "0.8"
-base64 = "0.21"
+base64 = "0.22"
+bytes = "1"
 cfg-if = "1.0"
 ctr = "0.9"
 dns-sd = { version = "0.1.3", optional = true }
@@ -18,7 +19,9 @@ form_urlencoded = "1.0"
 futures-core = "0.3"
 futures-util = "0.3"
 hmac = "0.12"
-hyper = { version = "0.14", features = ["http1", "server", "tcp"] }
+hyper = { version = "1.3", features = ["http1"] }
+hyper-util = { version = "0.1", features = ["server-auto", "server-graceful", "service"] }
+http-body-util = "0.1.1"
 libmdns = "0.8"
 log = "0.4"
 rand = "0.8"
