@@ -1,7 +1,7 @@
 [package]
 name = "librespot-playback"
 version = "0.5.0-dev"
-rust-version = "1.61"
+rust-version = "1.73"
 authors = ["Sasha Hilton <sashahilton00@gmail.com>"]
 description = "The audio playback logic for librespot"
 license = "MIT"
@@ -43,7 +43,7 @@ gstreamer-audio = { version = "0.22.0", optional = true }
 glib            = { version = "0.19.2", optional = true }
 
 # Rodio dependencies
-rodio           = { version = "0.17.1", optional = true, default-features = false }
+rodio           = { version = "0.18.1", optional = true, default-features = false }
 cpal            = { version = "0.15.1", optional = true }
 
 # Container and audio decoder
