@@ -1,6 +1,7 @@
 use std::collections::VecDeque;
 
-use hyper::{Body, Method, Request};
+use bytes::Bytes;
+use hyper::{Method, Request};
 use serde::Deserialize;
 
 use crate::Error;
@@ -85,7 +86,7 @@ impl ApResolver {
         let req = Request::builder()
             .method(Method::GET)
             .uri("https://apresolve.spotify.com/?type=accesspoint&type=dealer&type=spclient")
-            .body(Body::empty())?;
+            .body(Bytes::new())?;
 
         let body = self.session().http_client().request_body(req).await?;
         let data: ApResolveData = serde_json::from_slice(body.as_ref())?;
